requirements.txt
flask
google-generativeai
python-dotenv
pandas
scikit-learn
xgboost
tensorflow
